requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
python-json-logger
transformers
